finish moving components into new folder structure
