+++ client/components/invoices/NewInvoice.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+// import { useSelector, useDispatch } from 'react-redux'
+import { Button } from '@chakra-ui/react'
+
+export function NewInvoice() {
+  return (
+    <>
+      <Button colorScheme="teal">Create Invoice</Button>
+    </>
+  )
+}
